add css, link to github, clean up
