--- README.md
@@ -1,13 +1,0 @@
-# dash-covid19-ca-bay-area
-
-## View the project: 
-## https://dash-covid19-ca-bay-area.herokuapp.com/
-
-### Project to see trends in Covid-19 cases and maintain Python and web-development skills.
-### I started this project as a Juypter notebook but have since deployed it as a dash app.
-
-### Contra Costa County did not have a way to see Covid-19 data trends on a more granualar level 
-### (i.e. on a per city basis). However, they provide the data as a download in Excel format, so I
-### built a simple script to download the data each day (starting in Fall 2020). For Alameda county
-### the data was open source, but in March 2022 the data source is no longer available. 
-
